First implementation of VaultODVExporter, factored out VaultExporter
