+++ plugins/de/opendiabetes/vault/plugin/exporter/vaultodv/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package contains the VaultODVExporter plugin.
+ */
+package de.opendiabetes.vault.plugin.exporter.vaultodv;
