+++ plugins/de/opendiabetes/vault/plugin/exporter/vaultodv/plugin.properties
@@ -1,0 +1,5 @@
+plugin.class=de.opendiabetes.vault.plugin.exporter.vaultodv.VaultODVExporter
+plugin.dependencies=
+plugin.id=VaultODVExporter
+plugin.provider=OpenDiabetes
+plugin.version=0.0.1
