+++ properties/vaultodvexporter.properties
@@ -1,0 +1,3 @@
+periodRestriction=true
+periodRestrictionFrom=01/12/2016
+periodRestrictionTo=01/12/2017
